+++ tests/unit/asana/test_client.py
@@ -86,8 +86,9 @@ def fixture_project_test():
 @pytest.fixture(name='sections_in_project_test', scope='session')
 def fixture_sections_in_project_test(project_test):
     """
-    Creates some test sections and returns a list of them, each of which is the
-    dict of data that should match the 'data' element returned by the API.
+    Creates some test sections in the test project and returns a list of them,
+    each of which is the dict of data that should match the `data` element
+    returned by the API.
 
     Will delete the sections once done with all tests.
 
@@ -118,8 +119,60 @@ def fixture_sections_in_project_test(project_test):
 
     yield sect_data_list
 
-    for i in range(num_sects):
-        client.sections.delete_section(sect_data_list[i]['gid'])
+    for sect_data in sect_data_list:
+        client.sections.delete_section(sect_data['gid'])
+
+
+
+@pytest.fixture(name='tasks_in_project_and_utl_test', scope='session')
+def fixture_tasks_in_project_and_utl_test(project_test,
+        sections_in_project_test, sections_in_utl_test):
+    """
+    Creates some tasks in both the user task list (in the test workspace) and
+    the test project, and returns a list of them, each of which is the dict of
+    data that should match the `data` element returned by the API.  The tasks
+    in the user task list and the test project are the same tasks.
+
+    Will delete the tasks once done with all tests.
+
+    This is not being used with the autouse keyword so that, if running tests
+    that do not require this section fixture, they can run more optimally
+    without the need to needlessly create and delete this section.  (Also,
+    could not figure out how to get rid of all syntax and pylint errors).
+
+    ** Consumes 7 API calls. **
+    (API call count is 3*num_sects + 1)
+    """
+    # pylint: disable=no-member     # asana.Client dynamically adds attrs
+    num_tasks = 2
+    client = aclient._get_client()
+    me_data = aclient._get_me()
+
+    task_data_list = []
+    for _ in range(num_tasks):
+        task_name = tester_data._TASK_TEMPLATE.substitute({'tid': uuid.uuid4()})
+        params = {
+            'assignee': me_data['gid'],
+            'assignee_section': sections_in_utl_test[0]['gid'],
+            'name': task_name,
+            'projects': [
+                project_test['gid'],
+            ],
+        }
+        task_data = client.tasks.create_task(params)
+        task_data_list.append(task_data)
+
+        # No way to add project section at task creation, so need separate call
+        params = {
+            'task': task_data['gid'],
+        }
+        client.sections.add_task_for_section(sections_in_project_test[0]['gid'],
+                params)
+
+    yield task_data_list
+
+    for task_data in task_data_list:
+        client.tasks.delete_task(task_data['gid'])
 
 
 
@@ -210,6 +263,7 @@ def test_asana_error_handler(caplog):
     'get_section_gid_from_name',
     'get_user_task_list_gid',
     'get_section_gids_in_project_or_utl',
+    'get_tasks',
 ])
 def test_dec_usage_asana_error_handler(func_name):
     """
@@ -576,6 +630,104 @@ def test_get_section_gids_in_project_or_utl(monkeypatch, caplog, project_test,
             raise_asana_error)
     subtest_asana_error_handler_func(caplog, asana.error.InvalidRequestError, 0,
             aclient.get_section_gids_in_project_or_utl, project_test['gid'])
+
+
+
+@pytest.mark.asana_error_data.with_args(asana.error.NoAuthorizationError)
+def test_get_tasks(monkeypatch, caplog,        # pylint: disable=too-many-locals
+        project_test, sections_in_project_test, sections_in_utl_test,
+        tasks_in_project_and_utl_test, raise_asana_error):
+    """
+    Tests the `get_tasks()` method.
+
+    This does require the asana account be configured to support unit testing.
+    See CONTRIBUTING.md.
+
+    ** Consumes at least 4 API call. **
+    (varies depending on data size, but only 4 calls intended)
+
+    Raises:
+      (TesterNotInitializedError): If test workspace does not exist on asana
+        account tied to access token, will stop test.  User must create
+        manually per docs.
+    """
+    # pylint: disable=no-member     # asana.Client dynamically adds attrs
+    caplog.set_level(logging.ERROR)
+
+    try:
+        me_data = aclient._get_me()
+    except aclient.DataNotFoundError as ex:
+        # This is an error with the tester, not the module under test
+        raise TesterNotInitializedError('Cannot run unit tests: Must create a'
+                + f' workspace named "{tester_data._WORKSPACE}" in the asana'
+                + ' account tied to access token in .secrets.conf') from ex
+
+    ws_gid = aclient.get_workspace_gid_from_name(tester_data._WORKSPACE)
+
+    params = {
+        'assignee': me_data['gid'],
+        'workspace': ws_gid,
+    }
+    fields = [
+        'assignee_section',
+        'due_at',
+        'name',
+        'projects',
+    ]
+    tasks_found = aclient.get_tasks(params, fields)
+    # Filter/match list since other tests may have added more tasks to server
+    tasks_to_check = {}
+    # Nested fors must be in this order - tasks_found is a single-iter generator
+    for task_found in tasks_found:
+        for i_task_expected, task_expected in enumerate(
+                tasks_in_project_and_utl_test):
+            if task_found['gid'] == task_expected['gid']:
+                tasks_to_check[i_task_expected] = task_found
+                break
+    assert len(tasks_to_check) == len(tasks_in_project_and_utl_test)
+    for i_task_expected, task_found in tasks_to_check.items():
+        task_expected = tasks_in_project_and_utl_test[i_task_expected]
+        assert task_found['assignee_section']['gid'] \
+                    == sections_in_utl_test[0]['gid']
+        assert task_found['due_at'] is None
+        assert task_found['name'] == task_expected['name']
+        assert task_found['projects'][0]['gid'] == project_test['gid']
+
+    params = {
+        'project': project_test['gid'],
+    }
+    fields = [
+        'due_on',
+        'memberships.section',
+        'name',
+        'projects',
+    ]
+    tasks_found = aclient.get_tasks(params, fields)
+    # Filter/match list since other tests may have added more tasks to server
+    tasks_to_check = {}
+    # Nested fors must be in this order - tasks_found is a single-iter generator
+    for task_found in tasks_found:
+        for i_task_expected, task_expected in enumerate(
+                tasks_in_project_and_utl_test):
+            if task_found['gid'] == task_expected['gid']:
+                tasks_to_check[i_task_expected] = task_found
+                break
+    assert len(tasks_to_check) == len(tasks_in_project_and_utl_test)
+    for i_task_expected, task_found in tasks_to_check.items():
+        task_expected = tasks_in_project_and_utl_test[i_task_expected]
+        assert task_found['due_on'] is None
+        assert task_found['name'] == task_expected['name']
+        assert task_found['projects'][0]['gid'] == project_test['gid']
+        assert sections_in_project_test[0]['gid'] in \
+                [m['section']['gid'] for m in task_found['memberships'] \
+                    if 'section' in m]
+
+    # Function-specific practical test of @asana_error_handler
+    client = aclient._get_client()
+    # Need to monkeypatch cached client since class dynamically creates attrs
+    monkeypatch.setattr(client.tasks, 'get_tasks', raise_asana_error)
+    subtest_asana_error_handler_func(caplog, asana.error.NoAuthorizationError,
+            0, aclient.get_tasks, {})